code-explainer/requirements.txt
dash
dash-bootstrap-components
databricks-sdk
python-dotenv
networkx
matplotlib
dash-cytoscape